Revert "Move transactions module"
This reverts commit 245e136a
