+++ docs/transactions.rst
@@ -1,13 +1,10 @@
 Transactions
 ============
 
-.. warning::
-    This module is deprecated and will be removed in the future versions.
-
-.. py:module:: starknet_py.net.transactions.deploy
+.. py:module:: starknet_py.transactions.deploy
 
 .. autofunction:: make_deploy_tx
 
-.. py:module:: starknet_py.net.transactions.declare
+.. py:module:: starknet_py.transactions.declare
 
 .. autofunction:: make_declare_tx
