+++ starknet_py/net/account/account_client.py
@@ -37,7 +37,7 @@ from starknet_py.net.models import (
 from starknet_py.net.networks import Network, MAINNET, TESTNET
 from starknet_py.net.signer.stark_curve_signer import StarkCurveSigner, KeyPair
 from starknet_py.net.signer import BaseSigner
-from starknet_py.net.transactions.deploy import make_deploy_tx
+from starknet_py.transactions.deploy import make_deploy_tx
 from starknet_py.utils.crypto.facade import Call
 from starknet_py.utils.data_transformer.execute_transformer import execute_transformer
 from starknet_py.utils.sync import add_sync_methods
